Harden app shell state and desktop access lanes
