+++ src/app/components/workspace/diff/diff-panel-file-list.tsx
@@ -9,6 +9,7 @@ import {
 } from '@pierre/diffs/react'
 import type { VirtualItem } from '@tanstack/react-virtual'
 import { ChevronDown, ChevronRight, MessageSquarePlus } from 'lucide-react'
+import { openPathQuery } from '../../../query/desktop-query'
 import {
   DIFF_PANEL_UNSAFE_CSS,
   type DiffCommentMetadata,
@@ -172,10 +173,7 @@ function DiffPanelFileRow({
       onPointerDownCapture={(event) => handleFilePointerDownCapture(event, fileKey, filePath)}
       onClickCapture={(event) => {
         if (!isDiffHeaderClick(event)) return
-        const openPathPromise = window.piDesktop?.openPath?.(
-          joinProjectFilePath(projectId, filePath),
-        )
-        void openPathPromise?.catch(() => undefined)
+        void openPathQuery(joinProjectFilePath(projectId, filePath)).catch(() => undefined)
       }}
     >
       <FileDiff<DiffCommentMetadata>
